+++ src/theory/quantifiers/ematching/inst_strategy_e_matching_user.h
@@ -21,6 +21,7 @@
 #include "expr/node.h"
 #include "theory/quantifiers/ematching/inst_strategy.h"
 #include "theory/quantifiers/ematching/trigger.h"
+#include "options/quantifiers_options.h"
 
 namespace CVC4 {
 namespace theory {
@@ -51,6 +52,11 @@ class InstStrategyUserPatterns : public InstStrategy
   void processResetInstantiationRound(Theory::Effort effort) override;
   /** Process quantified formula q at the given effort */
   InstStrategyStatus process(Node f, Theory::Effort effort, int e) override;
+  /** 
+   * Get the current user pat mode, which may be interleaved based on counters
+   * maintained by the quantifiers state.
+   */
+  options::UserPatMode getInstUserPatMode() const;
   /** explicitly provided patterns */
   std::map<Node, std::vector<inst::Trigger*> > d_user_gen;
   /** waiting to be generated patterns */
